+++ src/adjusters/numericString.mjs
@@ -7,6 +7,7 @@ import Convert from "../decorators/convert";
 import Type from "../decorators/string/type";
 import MinLength from "../decorators/string/minLength";
 import MaxLength from "../decorators/string/maxLength";
+import FullWidthToHalf from "../decorators/numericString/fullWidthToHalf";
 import JoinArray from "../decorators/numericString/joinArray";
 import SeparatedBy from "../decorators/numericString/separatedBy";
 import Pattern from "../decorators/numericString/pattern";
@@ -32,6 +33,7 @@ export default () =>
 @SeparatedBy
 @Type
 @JoinArray
+@FullWidthToHalf
 @AcceptEmptyString
 @AcceptNull
 @Default