+++ src/index.d.ts
@@ -122,6 +122,12 @@ interface NumberAdjuster extends AdjusterBase<number>
 	acceptSpecialFormats(): this
 
 	/**
+	 * accept full width numeric string; i.e., "１２３．４５６"
+	 * @returns chainable instance
+	 */
+	acceptFullWidth(): this
+
+	/**
 	 * limit value to integer chain
 	 * @param [adjust=false] adjust to integer value is not an integer; default is ERROR
 	 * @returns chainable instance
@@ -269,6 +275,12 @@ interface NumericStringAdjuster extends AdjusterBase<string>
 	 * @returns chainable instance
 	 */
 	acceptEmptyString(value?: string | null): this
+
+	/**
+	 * convert full width string to half width
+	 * @returns chainable instance
+	 */
+	fullWidthToHalf(): this
 
 	/**
 	 * join array into string