+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* `adjuster.number().acceptFullWidth()`
+* `adjuster.numericString().fullWidthToHalf()`
 * `adjuster.number().convert()`
 * `adjuster.string().convert()`
 * `adjuster.numericString().convert()`